feat(claw): redirect claw landing to chat
