--- apps/web/src/app/(app)/organizations/[id]/claw/OrgClawDashboardClient.tsx
@@ -1,16 +1,0 @@
-'use client';
-
-import { useOrgKiloClawStatus } from '@/hooks/useOrgKiloClaw';
-import { ClawDashboard } from '@/app/(app)/claw/components/ClawDashboard';
-import { withStatusQueryBoundary } from '@/app/(app)/claw/components/withStatusQueryBoundary';
-
-const ClawDashboardWithBoundary = withStatusQueryBoundary(ClawDashboard);
-
-function OrgClawDashboardLoader({ organizationId }: { organizationId: string }) {
-  const statusQuery = useOrgKiloClawStatus(organizationId);
-  return <ClawDashboardWithBoundary statusQuery={statusQuery} organizationId={organizationId} />;
-}
-
-export function OrgClawDashboardClient({ organizationId }: { organizationId: string }) {
-  return <OrgClawDashboardLoader organizationId={organizationId} />;
-}
