+++ apps/web/src/app/(app)/organizations/[id]/claw/page.tsx
@@ -1,5 +1,5 @@
 import { OrganizationByPageLayout } from '@/components/organizations/OrganizationByPageLayout';
-import { OrgClawDashboardClient } from './OrgClawDashboardClient';
+import { OrgClawRedirectClient } from './OrgClawRedirectClient';
 
 type OrgClawPageProps = {
   params: Promise<{ id: string }>;
@@ -9,7 +9,7 @@ export default async function OrgClawPage({ params }: OrgClawPageProps) {
   return (
     <OrganizationByPageLayout
       params={params}
-      render={org => <OrgClawDashboardClient organizationId={org.organization.id} />}
+      render={org => <OrgClawRedirectClient organizationId={org.organization.id} />}
     />
   );
 }
